Caverns Event Idea
This is the code I've been playing with to try to trigger an event when you enter a cavern. I think it may have some bugs, but Visual Studio wouldn't allow me to test just my code. So far I have no errors. I wanted the user to come into the caverns to find a sword.