Update Enemy.cs
I followed the guidelines to make the enemy able to attack and take damage. I also gave it a name, health, attack, and defend. I ran into some problems with the attack feature and other places where I couldn't figure out how to fix it. It may be in the wrong class or someone may have to make something somewhere else. I also added in a specific enemy type the Acid Worm and enabled to spit acid which causes poison damage over time to player.