Add files via upload
I added more to the intro of the game with the vibes of a spooky dungeon. I wasn't quite sure on what to do to improve the game, so I decided to add an alternate version to the intro of the game. There's not a whole lot that I added, but I do think it's still very effective.