Kitchen Return Code
For my task, I was asked to think of return prompts for the kitchen. The user comes back to the kitchen five times and each time finds a different clue and the clue gives them a better idea of what's in the dungeon with them. I was having issues with the code's compliance with main, and every time I tried to add or delete things to work out the error it just was not wanting to work. I'm not sure what's wrong because nothing I'm doing seems to be fixing. All my lines of code seem to be correct so I'm guessing it's a problem with another room.