Update Kitchen.cs
Updated with changes to the main branch.

Changed static void Main to Static void Kitchen on line 12. Cannot reuse Main. Main is defined in the Main program. All side rooms should be called their name. Integrating as this will not break the rest of the program now, but it still needs to be called/used in the main program area when the user enters the kitchen for each instance. There are two parts to make this work. The first part is this code you've made. The second part is using it within the main program.